+++ webnodeex/exception.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+class WebNodeException(Exception):
+    pass
